requirements.txt
asgiref==3.6.0
Django==3.2
django-grappelli==3.0.5
django-tinymce==3.6.1
isort==5.12.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
